+++ randomizer.py
@@ -364,8 +364,6 @@ class Randomizer:
       if self.options.get("sword_mode") == "Swordless":
         patcher.apply_patch(self, "swordless")
         tweaks.update_text_for_swordless(self)
-      if self.options.get("randomize_entrances") not in ["Disabled", None, "Dungeons"]:
-        tweaks.disable_ice_ring_isle_and_fire_mountain_effects_indoors(self)
       tweaks.update_starting_gear(self)
       if self.options.get("disable_tingle_chests_with_tingle_bombs"):
         patcher.apply_patch(self, "disable_tingle_bombs_on_tingle_chests")